requirements.txt
networkx==3.2.1        
pymongo==4.6.3         
faiss-cpu==1.8.0       

google-generativeai==0.4.1  

numpy==1.26.4          
pandas==2.2.1          
pyyaml==6.0.1       
dataclasses==0.8
typing==3.10
